+++ prover/crates/bin/witness_generator/src/rounds/leaf_aggregation/job_processor.rs
@@ -10,51 +10,48 @@ use zksync_types::basic_fri_types::AggregationRound;
 use crate::{
     artifacts::ArtifactsManager,
     metrics::WITNESS_GENERATOR_METRICS,
-    node_aggregation::{
-        NodeAggregationArtifacts, NodeAggregationWitnessGenerator,
-        NodeAggregationWitnessGeneratorJob,
+    rounds::{
+        leaf_aggregation::{
+            LeafAggregation, LeafAggregationArtifacts, LeafAggregationWitnessGeneratorJob,
+        },
+        JobManager, WitnessGenerator,
     },
-    witness_generator::WitnessGenerator,
 };
 
 #[async_trait]
-impl JobProcessor for NodeAggregationWitnessGenerator {
-    type Job = NodeAggregationWitnessGeneratorJob;
+impl JobProcessor for WitnessGenerator<LeafAggregation> {
+    type Job = LeafAggregationWitnessGeneratorJob;
     type JobId = u32;
-    type JobArtifacts = NodeAggregationArtifacts;
+    type JobArtifacts = LeafAggregationArtifacts;
 
-    const SERVICE_NAME: &'static str = "fri_node_aggregation_witness_generator";
+    const SERVICE_NAME: &'static str = "fri_leaf_aggregation_witness_generator";
 
     async fn get_next_job(&self) -> anyhow::Result<Option<(Self::JobId, Self::Job)>> {
-        let mut prover_connection = self.prover_connection_pool.connection().await?;
+        let mut prover_connection = self.connection_pool.connection().await?;
         let pod_name = get_current_pod_name();
         let Some(metadata) = prover_connection
             .fri_witness_generator_dal()
-            .get_next_node_aggregation_job(self.protocol_version, &pod_name)
+            .get_next_leaf_aggregation_job(self.protocol_version, &pod_name)
             .await
         else {
             return Ok(None);
         };
-        tracing::info!("Processing node aggregation job {:?}", metadata.id);
+        tracing::info!("Processing leaf aggregation job {:?}", metadata.id);
         Ok(Some((
             metadata.id,
-            <Self as WitnessGenerator>::prepare_job(
-                metadata,
-                &*self.object_store,
-                self.keystore.clone(),
-            )
-            .await
-            .context("prepare_job()")?,
+            <Self as JobManager>::prepare_job(metadata, &*self.object_store, self.keystore.clone())
+                .await
+                .context("prepare_leaf_aggregation_job()")?,
         )))
     }
 
     async fn save_failure(&self, job_id: u32, _started_at: Instant, error: String) -> () {
-        self.prover_connection_pool
+        self.connection_pool
             .connection()
             .await
             .unwrap()
             .fri_witness_generator_dal()
-            .mark_node_aggregation_job_failed(&error, job_id)
+            .mark_leaf_aggregation_job_failed(&error, job_id)
             .await;
     }
 
@@ -62,13 +59,13 @@ impl JobProcessor for NodeAggregationWitnessGenerator {
     async fn process_job(
         &self,
         _job_id: &Self::JobId,
-        job: NodeAggregationWitnessGeneratorJob,
+        job: LeafAggregationWitnessGeneratorJob,
         started_at: Instant,
-    ) -> tokio::task::JoinHandle<anyhow::Result<NodeAggregationArtifacts>> {
+    ) -> tokio::task::JoinHandle<anyhow::Result<LeafAggregationArtifacts>> {
         let object_store = self.object_store.clone();
         let max_circuits_in_flight = self.config.max_circuits_in_flight;
         tokio::spawn(async move {
-            <Self as WitnessGenerator>::process_job(
+            <Self as JobManager>::process_job(
                 job,
                 object_store,
                 Some(max_circuits_in_flight),
@@ -78,25 +75,34 @@ impl JobProcessor for NodeAggregationWitnessGenerator {
         })
     }
 
-    #[tracing::instrument(
-        skip_all,
-        fields(l1_batch = % artifacts.block_number, circuit_id = % artifacts.circuit_id)
-    )]
     async fn save_result(
         &self,
         job_id: u32,
         started_at: Instant,
-        artifacts: NodeAggregationArtifacts,
+        artifacts: LeafAggregationArtifacts,
     ) -> anyhow::Result<()> {
+        let block_number = artifacts.block_number;
+        let circuit_id = artifacts.circuit_id;
+        tracing::info!(
+            "Saving leaf aggregation artifacts for block {} with circuit {}",
+            block_number.0,
+            circuit_id,
+        );
+
         let blob_save_started_at = Instant::now();
 
         let blob_urls = Self::save_to_bucket(job_id, artifacts.clone(), &*self.object_store).await;
 
-        WITNESS_GENERATOR_METRICS.blob_save_time[&AggregationRound::NodeAggregation.into()]
+        WITNESS_GENERATOR_METRICS.blob_save_time[&AggregationRound::LeafAggregation.into()]
             .observe(blob_save_started_at.elapsed());
 
+        tracing::info!(
+            "Saved leaf aggregation artifacts for block {} with circuit {}",
+            block_number.0,
+            circuit_id,
+        );
         Self::save_to_database(
-            &self.prover_connection_pool,
+            &self.connection_pool,
             job_id,
             started_at,
             blob_urls,
@@ -112,15 +118,15 @@ impl JobProcessor for NodeAggregationWitnessGenerator {
 
     async fn get_job_attempts(&self, job_id: &u32) -> anyhow::Result<u32> {
         let mut prover_storage = self
-            .prover_connection_pool
+            .connection_pool
             .connection()
             .await
-            .context("failed to acquire DB connection for NodeAggregationWitnessGenerator")?;
+            .context("failed to acquire DB connection for LeafAggregationWitnessGenerator")?;
         prover_storage
             .fri_witness_generator_dal()
-            .get_node_aggregation_job_attempts(*job_id)
+            .get_leaf_aggregation_job_attempts(*job_id)
             .await
             .map(|attempts| attempts.unwrap_or(0))
-            .context("failed to get job attempts for NodeAggregationWitnessGenerator")
+            .context("failed to get job attempts for LeafAggregationWitnessGenerator")
     }
 }
