+++ prover/crates/bin/witness_generator/src/rounds/node_aggregation/job_processor.rs
@@ -9,52 +9,49 @@ use zksync_types::basic_fri_types::AggregationRound;
 
 use crate::{
     artifacts::ArtifactsManager,
-    leaf_aggregation::{
-        LeafAggregationArtifacts, LeafAggregationWitnessGenerator,
-        LeafAggregationWitnessGeneratorJob,
-    },
     metrics::WITNESS_GENERATOR_METRICS,
-    witness_generator::WitnessGenerator,
+    rounds::{
+        node_aggregation::{
+            NodeAggregation, NodeAggregationArtifacts, NodeAggregationWitnessGeneratorJob,
+        },
+        JobManager, WitnessGenerator,
+    },
 };
 
 #[async_trait]
-impl JobProcessor for LeafAggregationWitnessGenerator {
-    type Job = LeafAggregationWitnessGeneratorJob;
+impl JobProcessor for WitnessGenerator<NodeAggregation> {
+    type Job = NodeAggregationWitnessGeneratorJob;
     type JobId = u32;
-    type JobArtifacts = LeafAggregationArtifacts;
+    type JobArtifacts = NodeAggregationArtifacts;
 
-    const SERVICE_NAME: &'static str = "fri_leaf_aggregation_witness_generator";
+    const SERVICE_NAME: &'static str = "fri_node_aggregation_witness_generator";
 
     async fn get_next_job(&self) -> anyhow::Result<Option<(Self::JobId, Self::Job)>> {
-        let mut prover_connection = self.prover_connection_pool.connection().await?;
+        let mut prover_connection = self.connection_pool.connection().await?;
         let pod_name = get_current_pod_name();
         let Some(metadata) = prover_connection
             .fri_witness_generator_dal()
-            .get_next_leaf_aggregation_job(self.protocol_version, &pod_name)
+            .get_next_node_aggregation_job(self.protocol_version, &pod_name)
             .await
         else {
             return Ok(None);
         };
-        tracing::info!("Processing leaf aggregation job {:?}", metadata.id);
+        tracing::info!("Processing node aggregation job {:?}", metadata.id);
         Ok(Some((
             metadata.id,
-            <Self as WitnessGenerator>::prepare_job(
-                metadata,
-                &*self.object_store,
-                self.keystore.clone(),
-            )
-            .await
-            .context("prepare_leaf_aggregation_job()")?,
+            <Self as JobManager>::prepare_job(metadata, &*self.object_store, self.keystore.clone())
+                .await
+                .context("prepare_job()")?,
         )))
     }
 
     async fn save_failure(&self, job_id: u32, _started_at: Instant, error: String) -> () {
-        self.prover_connection_pool
+        self.connection_pool
             .connection()
             .await
             .unwrap()
             .fri_witness_generator_dal()
-            .mark_leaf_aggregation_job_failed(&error, job_id)
+            .mark_node_aggregation_job_failed(&error, job_id)
             .await;
     }
 
@@ -62,13 +59,13 @@ impl JobProcessor for LeafAggregationWitnessGenerator {
     async fn process_job(
         &self,
         _job_id: &Self::JobId,
-        job: LeafAggregationWitnessGeneratorJob,
+        job: NodeAggregationWitnessGeneratorJob,
         started_at: Instant,
-    ) -> tokio::task::JoinHandle<anyhow::Result<LeafAggregationArtifacts>> {
+    ) -> tokio::task::JoinHandle<anyhow::Result<NodeAggregationArtifacts>> {
         let object_store = self.object_store.clone();
         let max_circuits_in_flight = self.config.max_circuits_in_flight;
         tokio::spawn(async move {
-            <Self as WitnessGenerator>::process_job(
+            <Self as JobManager>::process_job(
                 job,
                 object_store,
                 Some(max_circuits_in_flight),
@@ -78,34 +75,25 @@ impl JobProcessor for LeafAggregationWitnessGenerator {
         })
     }
 
+    #[tracing::instrument(
+        skip_all,
+        fields(l1_batch = % artifacts.block_number, circuit_id = % artifacts.circuit_id)
+    )]
     async fn save_result(
         &self,
         job_id: u32,
         started_at: Instant,
-        artifacts: LeafAggregationArtifacts,
+        artifacts: NodeAggregationArtifacts,
     ) -> anyhow::Result<()> {
-        let block_number = artifacts.block_number;
-        let circuit_id = artifacts.circuit_id;
-        tracing::info!(
-            "Saving leaf aggregation artifacts for block {} with circuit {}",
-            block_number.0,
-            circuit_id,
-        );
-
         let blob_save_started_at = Instant::now();
 
         let blob_urls = Self::save_to_bucket(job_id, artifacts.clone(), &*self.object_store).await;
 
-        WITNESS_GENERATOR_METRICS.blob_save_time[&AggregationRound::LeafAggregation.into()]
+        WITNESS_GENERATOR_METRICS.blob_save_time[&AggregationRound::NodeAggregation.into()]
             .observe(blob_save_started_at.elapsed());
 
-        tracing::info!(
-            "Saved leaf aggregation artifacts for block {} with circuit {}",
-            block_number.0,
-            circuit_id,
-        );
-        Self::save_to_database(
-            &self.prover_connection_pool,
+        <Self as ArtifactsManager>::save_to_database(
+            &self.connection_pool,
             job_id,
             started_at,
             blob_urls,
@@ -121,15 +109,15 @@ impl JobProcessor for LeafAggregationWitnessGenerator {
 
     async fn get_job_attempts(&self, job_id: &u32) -> anyhow::Result<u32> {
         let mut prover_storage = self
-            .prover_connection_pool
+            .connection_pool
             .connection()
             .await
-            .context("failed to acquire DB connection for LeafAggregationWitnessGenerator")?;
+            .context("failed to acquire DB connection for NodeAggregationWitnessGenerator")?;
         prover_storage
             .fri_witness_generator_dal()
-            .get_leaf_aggregation_job_attempts(*job_id)
+            .get_node_aggregation_job_attempts(*job_id)
             .await
             .map(|attempts| attempts.unwrap_or(0))
-            .context("failed to get job attempts for LeafAggregationWitnessGenerator")
+            .context("failed to get job attempts for NodeAggregationWitnessGenerator")
     }
 }
