+++ prover/crates/bin/witness_generator/src/rounds/scheduler/job_processor.rs
@@ -10,50 +10,48 @@ use zksync_types::{basic_fri_types::AggregationRound, L1BatchNumber};
 use crate::{
     artifacts::ArtifactsManager,
     metrics::WITNESS_GENERATOR_METRICS,
-    recursion_tip::{
-        RecursionTipArtifacts, RecursionTipJobMetadata, RecursionTipWitnessGenerator,
-        RecursionTipWitnessGeneratorJob,
+    rounds::{
+        scheduler::{
+            Scheduler, SchedulerArtifacts, SchedulerWitnessGeneratorJob,
+            SchedulerWitnessJobMetadata,
+        },
+        JobManager, WitnessGenerator,
     },
-    witness_generator::WitnessGenerator,
 };
 
 #[async_trait]
-impl JobProcessor for RecursionTipWitnessGenerator {
-    type Job = RecursionTipWitnessGeneratorJob;
+impl JobProcessor for WitnessGenerator<Scheduler> {
+    type Job = SchedulerWitnessGeneratorJob;
     type JobId = L1BatchNumber;
-    type JobArtifacts = RecursionTipArtifacts;
+    type JobArtifacts = SchedulerArtifacts;
 
-    const SERVICE_NAME: &'static str = "recursion_tip_witness_generator";
+    const SERVICE_NAME: &'static str = "fri_scheduler_witness_generator";
 
     async fn get_next_job(&self) -> anyhow::Result<Option<(Self::JobId, Self::Job)>> {
-        let mut prover_connection = self.prover_connection_pool.connection().await?;
+        let mut prover_connection = self.connection_pool.connection().await?;
         let pod_name = get_current_pod_name();
-        let Some((l1_batch_number, number_of_final_node_jobs)) = prover_connection
+        let Some(l1_batch_number) = prover_connection
             .fri_witness_generator_dal()
-            .get_next_recursion_tip_witness_job(self.protocol_version, &pod_name)
+            .get_next_scheduler_witness_job(self.protocol_version, &pod_name)
             .await
         else {
             return Ok(None);
         };
-
-        let final_node_proof_job_ids = prover_connection
+        let recursion_tip_job_id = prover_connection
             .fri_prover_jobs_dal()
-            .get_final_node_proof_job_ids_for(l1_batch_number)
-            .await;
-
-        assert_eq!(
-            final_node_proof_job_ids.len(),
-            number_of_final_node_jobs as usize,
-            "recursion tip witness job was scheduled without all final node jobs being completed; expected {}, got {}",
-            number_of_final_node_jobs, final_node_proof_job_ids.len()
-        );
+            .get_recursion_tip_proof_job_id(l1_batch_number)
+            .await
+            .context(format!(
+                "could not find recursion tip proof for l1 batch {}",
+                l1_batch_number
+            ))?;
 
         Ok(Some((
             l1_batch_number,
-            <Self as WitnessGenerator>::prepare_job(
-                RecursionTipJobMetadata {
+            <Self as JobManager>::prepare_job(
+                SchedulerWitnessJobMetadata {
                     l1_batch_number,
-                    final_node_proof_job_ids,
+                    recursion_tip_job_id,
                 },
                 &*self.object_store,
                 self.keystore.clone(),
@@ -64,12 +62,12 @@ impl JobProcessor for RecursionTipWitnessGenerator {
     }
 
     async fn save_failure(&self, job_id: L1BatchNumber, _started_at: Instant, error: String) -> () {
-        self.prover_connection_pool
+        self.connection_pool
             .connection()
             .await
             .unwrap()
             .fri_witness_generator_dal()
-            .mark_recursion_tip_job_failed(&error, job_id)
+            .mark_scheduler_job_failed(&error, job_id)
             .await;
     }
 
@@ -77,12 +75,12 @@ impl JobProcessor for RecursionTipWitnessGenerator {
     async fn process_job(
         &self,
         _job_id: &Self::JobId,
-        job: RecursionTipWitnessGeneratorJob,
+        job: SchedulerWitnessGeneratorJob,
         started_at: Instant,
-    ) -> tokio::task::JoinHandle<anyhow::Result<RecursionTipArtifacts>> {
+    ) -> tokio::task::JoinHandle<anyhow::Result<SchedulerArtifacts>> {
         let object_store = self.object_store.clone();
         tokio::spawn(async move {
-            <Self as WitnessGenerator>::process_job(job, object_store, None, started_at).await
+            <Self as JobManager>::process_job(job, object_store, None, started_at).await
         })
     }
 
@@ -94,18 +92,18 @@ impl JobProcessor for RecursionTipWitnessGenerator {
         &self,
         job_id: L1BatchNumber,
         started_at: Instant,
-        artifacts: RecursionTipArtifacts,
+        artifacts: SchedulerArtifacts,
     ) -> anyhow::Result<()> {
         let blob_save_started_at = Instant::now();
 
         let blob_urls =
             Self::save_to_bucket(job_id.0, artifacts.clone(), &*self.object_store).await;
 
-        WITNESS_GENERATOR_METRICS.blob_save_time[&AggregationRound::RecursionTip.into()]
+        WITNESS_GENERATOR_METRICS.blob_save_time[&AggregationRound::Scheduler.into()]
             .observe(blob_save_started_at.elapsed());
 
         Self::save_to_database(
-            &self.prover_connection_pool,
+            &self.connection_pool,
             job_id.0,
             started_at,
             blob_urls,
@@ -122,15 +120,15 @@ impl JobProcessor for RecursionTipWitnessGenerator {
 
     async fn get_job_attempts(&self, job_id: &L1BatchNumber) -> anyhow::Result<u32> {
         let mut prover_storage = self
-            .prover_connection_pool
+            .connection_pool
             .connection()
             .await
-            .context("failed to acquire DB connection for RecursionTipWitnessGenerator")?;
+            .context("failed to acquire DB connection for SchedulerWitnessGenerator")?;
         prover_storage
             .fri_witness_generator_dal()
-            .get_recursion_tip_witness_job_attempts(*job_id)
+            .get_scheduler_witness_job_attempts(*job_id)
             .await
             .map(|attempts| attempts.unwrap_or(0))
-            .context("failed to get job attempts for RecursionTipWitnessGenerator")
+            .context("failed to get job attempts for SchedulerWitnessGenerator")
     }
 }
